Updating the CSS and Layout,
Made changes to make the site more adaptive so it can work on phones

+++ deleteUser.php
@@ -5,11 +5,17 @@ require_once "ConnectDB.php";
 
 try{
     $userID = $_POST['Xdata'];
-    
-    $qry = "DELETE FROM users WHERE ID = ?;";
-    $stmt = $conn->prepare($qry);
-    $stmt->execute([$userID]);
 
+
+    if (trim($_SESSION['UserID']) == trim($userID)){
+        $msg = "<p><b class = 'error'>No Matter How Bad You Feel You Never Want To Delete Your Self!</b></p>";
+        $_SESSION['msg'] = $msg;
+
+    }else{
+        $qry = "DELETE FROM users WHERE ID = ?;";
+        $stmt = $conn->prepare($qry);
+        $stmt->execute([$userID]);
+    }
     header("location: manageUsers.php");
 } catch (PDOException $e) {
     echo "Error : ".$e->getMessage();
